requirements.txt
discord-py-interactions
morshutalk
